Add files via upload
Sorry for multiple commits - this new folder git_src contains all you need for running the markov_model and storing the generated midifiles in the 'generated_MIDI' folder that is in the 'model_weights_generatedMIDI.zip' file, this zip file also contains the folders model and weights that have to be extract on git_src folder, as well as the Data.zip folder. I hope it is clear enough. 
LSTM input still does not run from main -